(PC-18695) feat(bookings): move EndedBookingsSection from page to components
