+++ src/features/bookings/components/OnGoingBookingsList.tsx
@@ -4,7 +4,7 @@ import { useSafeAreaInsets } from 'react-native-safe-area-context'
 import styled from 'styled-components/native'
 
 import { useBookings } from 'features/bookings/api'
-import { EndedBookingsSection } from 'features/bookings/pages/EndedBookingsSection'
+import { EndedBookingsSection } from 'features/bookings/components/EndedBookingsSection'
 import { Booking } from 'features/bookings/types'
 import { analytics, isCloseToBottom } from 'libs/firebase/analytics'
 import useFunctionOnce from 'libs/hooks/useFunctionOnce'
